refactor(api): Move async func to abstract class
Remove the async function implementation from all other
classes and have it only in the abstruct class. If someone wants to
override this, it still possible in the child classe

Remove the async function implementation from all other
classes and have it only in the abstruct class. If someone wants to
override this, it still possible in the child classes.