feat(api): add support to print struct as FString
This commit adds support to print a struct as FString. With the
current implementation the string is formatted as JSON. This is
useful for debugging purposes.
However, the format may change in the future so it is important
to rely on the output being JSON.
It is only available if the feature `apigear` is enabled.

This is closes #75.