feat(msgbus): enable heartbeat on service side
The service tracks the addresses and last heartbeats for each client.
If the client missed to ping twice the heartbeat frequency it is assumed
that the connection is lost and the client is removed from the connected
clients list. This mean it will not get any property updates or signals
anymore unless it reconnects.