fix(msgbus): reset sent prop state on new value
When the property state was changed outside of the client and
accidentally set to the old cached sent value, the new set request was
not sent anymore.
This commit updates the sent cache when a new property state is set from
outside the client so that it can be requested again.
This also adds some test cases to cover this setup.

(cherry picked from commit 45c15e5e99be336907ad86a9ce2daa928a9d1f65)